Moved helper function out.
