+++ src/features/vote/ActiveRequests.tsx
@@ -5,7 +5,7 @@ import timerSVG from "assets/icons/timer.svg";
 import ActiveRequestsForm from "./ActiveRequestsForm";
 import { usePendingRequests, useVotePhase } from "hooks";
 import { OnboardContext } from "common/context/OnboardContext";
-import { DateTime } from "luxon";
+// import { DateTime } from "luxon";
 
 interface Props {
   // activeRequests: PriceRound[];
